django_csv/requirements.txt
amqp==5.3.1
asgiref==3.8.1
atpublic==5.1
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.7
django-celery-results==2.5.1
djangorestframework==3.15.2
flower==2.0.1
gunicorn==23.0.0
humanize==4.12.1
kombu==5.5.0
packaging==24.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
